refactor: 랜딩페이지 컴포넌트 코드 분리 #41
